Combine Serializer and Unserializer into a single interface/class
