+++ src/Serialization/Serializer.php
@@ -10,12 +10,18 @@ use Prezly\Slate\Model\Leaf;
 use Prezly\Slate\Model\Mark;
 use Prezly\Slate\Model\Text;
 use Prezly\Slate\Model\Value;
+use RuntimeException;
 
 class Serializer implements ValueSerializer
 {
     public function toJson(Value $value, int $json_options = null): string
     {
         return json_encode($this->serializeEntity($value), $json_options ?? 0);
+    }
+
+    public function fromJson(string $value): Value
+    {
+        return $this->unserializeValue(json_decode($value, false));
     }
 
     private function serializeEntity(Entity $entity): array
@@ -44,12 +50,56 @@ class Serializer implements ValueSerializer
         throw new InvalidArgumentException('Unsupported entity type given: ' . get_class($entity));
     }
 
+    /**
+     * @param mixed $object
+     * @return Object|Block|Inline|Text
+     */
+    private function unserializeEntity($object): Entity
+    {
+        $object = ShapeValidator::validateSlateObject($object); // generic slate object check
+
+        switch ($object->object) {
+            case Entity::VALUE:
+                return $this->unserializeValue($object);
+
+            case Entity::DOCUMENT:
+                return $this->unserializeDocument($object);
+
+            case Entity::BLOCK:
+                return $this->unserializeBlock($object);
+
+            case Entity::INLINE:
+                return $this->unserializeInline($object);
+
+            case Entity::TEXT:
+                return $this->unserializeText($object);
+
+            case Entity::LEAF:
+                return $this->unserializeLeaf($object);
+
+            case Entity::MARK:
+                return $this->unserializeMark($object);
+        }
+        throw new RuntimeException("Unsupported object type: " . $object->object);
+    }
+
     private function serializeValue(Value $value): array
     {
         return [
             'object'   => Entity::VALUE,
             'document' => $this->serializeDocument($value->getDocument()),
         ];
+    }
+
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Value
+     */
+    private function unserializeValue($object): Value
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::VALUE, ['document' => 'is_object']);
+
+        return new Value($this->unserializeDocument($object->document));
     }
 
     private function serializeDocument(Document $document): array
@@ -61,6 +111,25 @@ class Serializer implements ValueSerializer
                 return $this->serializeBlock($block);
             }, $document->getNodes()),
         ];
+    }
+
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Document
+     */
+    private function unserializeDocument($object): Document
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::DOCUMENT, [
+            'data'  => 'is_object',
+            'nodes' => 'is_array',
+        ]);
+
+        $nodes = [];
+        foreach ($object->nodes as $node) {
+            $nodes[] = $this->unserializeBlock($node);
+        }
+
+        return new Document($nodes, (array) $object->data);
     }
 
     private function serializeBlock(Block $block): array
@@ -75,6 +144,22 @@ class Serializer implements ValueSerializer
         ];
     }
 
+    private function unserializeBlock($object): Block
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::BLOCK, [
+            'type'  => 'is_string',
+            'data'  => 'is_object',
+            'nodes' => 'is_array',
+        ]);
+
+        $nodes = [];
+        foreach ($object->nodes as $node) {
+            $nodes[] = $this->unserializeEntity($node);
+        }
+
+        return new Block($object->type, $nodes, (array) $object->data);
+    }
+
     private function serializeInline(Inline $inline): array
     {
         return [
@@ -87,6 +172,26 @@ class Serializer implements ValueSerializer
         ];
     }
 
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Inline
+     */
+    private function unserializeInline($object): Inline
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::INLINE, [
+            'type'  => 'is_string',
+            'data'  => 'is_object',
+            'nodes' => 'is_array',
+        ]);
+
+        $nodes = [];
+        foreach ($object->nodes as $node) {
+            $nodes[] = $this->unserializeEntity($node);
+        }
+
+        return new Inline($object->type, $nodes, (array) $object->data);
+    }
+
     private function serializeText(Text $text): array
     {
         return [
@@ -95,6 +200,24 @@ class Serializer implements ValueSerializer
                 return $this->serializeLeaf($leaf);
             }, $text->getLeaves())
         ];
+    }
+
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Text
+     */
+    private function unserializeText($object): Text
+    {
+        ShapeValidator::validateSlateObject($object, Entity::TEXT, [
+            'leaves' => 'is_array',
+        ]);
+
+        $leaves = [];
+        foreach ($object->leaves as $leaf) {
+            $leaves[] = $this->unserializeLeaf($leaf);
+        }
+
+        return new Text($leaves);
     }
 
     private function serializeLeaf(Leaf $leaf): array
@@ -108,6 +231,25 @@ class Serializer implements ValueSerializer
         ];
     }
 
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Leaf
+     */
+    private function unserializeLeaf($object): Leaf
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::LEAF, [
+            'text'  => 'is_string',
+            'marks' => 'is_array',
+        ]);
+
+        $marks = [];
+        foreach ($object->marks as $mark) {
+            $marks[] = $this->unserializeMark($mark);
+        }
+
+        return new Leaf($object->text, $marks);
+    }
+
     private function serializeMark(Mark $mark): array
     {
         return [
@@ -115,5 +257,18 @@ class Serializer implements ValueSerializer
             'type'   => $mark->getType(),
             'data'   => (object) $mark->getData(),
         ];
+    }
+
+    /**
+     * @param mixed $object
+     * @return \Prezly\Slate\Model\Mark
+     */
+    private function unserializeMark($object): Mark
+    {
+        $object = ShapeValidator::validateSlateObject($object, Entity::MARK, [
+            'type' => 'is_string',
+            'data' => 'is_object',
+        ]);
+        return new Mark($object->type, (array) $object->data);
     }
 }
